# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Biblioteca de Prompts API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List prompts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private, all]
        - name: favorites
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptListResponse"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/popular:
    get:
      operationId: listPopularPrompts
      tags: [prompts]
      summary: List popular/trending public prompts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Popular prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
  /prompts/recent:
    get:
      operationId: listRecentPrompts
      tags: [prompts]
      summary: List recently created public prompts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prompt"
  /prompts/{id}:
    get:
      operationId: getPrompt
      tags: [prompts]
      summary: Get a prompt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update a prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePrompt
      tags: [prompts]
      summary: Delete a prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /prompts/{id}/favorite:
    patch:
      operationId: toggleFavorite
      tags: [prompts]
      summary: Toggle favorite on a prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}/like:
    patch:
      operationId: toggleLike
      tags: [prompts]
      summary: Toggle like on a public prompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}/use:
    post:
      operationId: recordUse
      tags: [prompts]
      summary: Record that a prompt was used (increments use count)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"

  # ─── AI ──────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Prompt:
      type: object
      required: [id, title, content, categoryId, categoryName, visibility, isFavorite, likeCount, useCount, rating, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        tags:
          type: array
          items:
            type: string
        visibility:
          type: string
          enum: [public, private]
        isFavorite:
          type: boolean
        likeCount:
          type: integer
        useCount:
          type: integer
        rating:
          type: number
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PromptInput:
      type: object
      required: [title, content, categoryId]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        description:
          type: string
        categoryId:
          type: integer
        tags:
          type: array
          items:
            type: string
        visibility:
          type: string
          enum: [public, private]
    PromptUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        description:
          type: string
        categoryId:
          type: integer
        tags:
          type: array
          items:
            type: string
        visibility:
          type: string
          enum: [public, private]
    PromptListResponse:
      type: object
      required: [prompts, total, page, limit]
      properties:
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/Prompt"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── AI ────────────────────────────────────────────────────────────────────
...