# lib/api-zod/src/generated/types/promptInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Biblioteca de Prompts API
 * OpenAPI spec version: 0.1.0
 */
import type { PromptInputVisibility } from './promptInputVisibility';

export interface PromptInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  description?: string;
  categoryId: number;
  tags?: string[];
  visibility?: PromptInputVisibility;
}